+++ .env.example
@@ -5,3 +5,5 @@ NAVER_APP_PASSWORD=
 
 # Optional
 # NAVER_CALDAV_CALENDAR=
+# NAVER_CALDAV_TIMEOUT_SECONDS=15
+# NAVER_CALDAV_INCLUDE_ALL_TODOS=false
