+++ src/yule_orchestrator/integrations/calendar/__init__.py
@@ -1,13 +1,6 @@
-from .naver_caldav import (
-    CalendarEvent,
-    CalendarIntegrationError,
-    CalendarQueryResult,
-    CalendarTodo,
-    list_naver_calendar_items,
-    list_naver_calendar_events,
-    render_calendar_items,
-    render_calendar_events,
-)
+from .models import CalendarEvent, CalendarQueryResult, CalendarTodo
+from .naver_caldav import CalendarIntegrationError, list_naver_calendar_events, list_naver_calendar_items
+from .rendering import render_calendar_events, render_calendar_items
 
 __all__ = [
     "CalendarEvent",
